解决编译问题

+++ interfaces/kits/components/ui_canvas.h
@@ -808,7 +808,7 @@ public:
         rotateAngle_ += angle;
         changeFlage_ = IsTransform();
     }
-    double GetRotate() const
+    float GetRotate() const
     {
         return rotateAngle_;
     }
@@ -901,7 +901,7 @@ private:
     double scaleY_;                          //y轴方向放大或缩小倍数
     double shearX_;                          //水平倾斜绘图
     double shearY_;                          //垂直倾斜绘图
-    double rotateAngle_;                     //旋转角度，单位度数
+    float rotateAngle_;                     //旋转角度，单位度数
     int16_t transLateX_;                     //X轴方向偏移像素
     int16_t transLateY_;                     //Y轴方向偏移像素
 };
